backend/pyproject.toml
[project]
name = "wecare-ai-tickets-backend"
version = "0.1.0"
description = "API FastAPI para tickets de TI com IA, Linear e e-mail."
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "fastapi>=0.115.6",
    "google-cloud-aiplatform>=1.85.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.6.1",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.32.1",
    "pyjwt[crypto]>=2.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]

[tool.uv]
default-groups = ["dev"]
